+++ addons/GameGUI/GGComponent.gd
@@ -62,7 +62,7 @@ enum TextSizeMode
 		if horizontal_mode == value: return
 		horizontal_mode = value
 		if value in [GGComponent.ScalingMode.ASPECT_FIT,GGComponent.ScalingMode.ASPECT_FILL]:
-			if not (vertical_mode in [value,GGComponent.ScalingMode.EXPAND_TO_FILL,GGComponent.ScalingMode.SHRINK_TO_FIT,GGComponent.ScalingMode.PARAMETER]): vertical_mode = value
+			if vertical_mode in [GGComponent.ScalingMode.PROPORTIONAL,GGComponent.ScalingMode.FIXED,GGComponent.ScalingMode.PARAMETER]: vertical_mode = value
 			if layout_size.x  < 0.0001: layout_size.x = 1
 			if layout_size.y  < 0.0001: layout_size.y = 1
 		elif vertical_mode in [GGComponent.ScalingMode.ASPECT_FIT,GGComponent.ScalingMode.ASPECT_FILL]:
@@ -78,7 +78,7 @@ enum TextSizeMode
 		if vertical_mode == value: return
 		vertical_mode = value
 		if value in [GGComponent.ScalingMode.ASPECT_FIT,GGComponent.ScalingMode.ASPECT_FILL]:
-			if not (horizontal_mode in [value,GGComponent.ScalingMode.EXPAND_TO_FILL,GGComponent.ScalingMode.SHRINK_TO_FIT,GGComponent.ScalingMode.PARAMETER]): horizontal_mode = value
+			if horizontal_mode in [GGComponent.ScalingMode.PROPORTIONAL,GGComponent.ScalingMode.FIXED,GGComponent.ScalingMode.PARAMETER]: horizontal_mode = value
 			if abs(layout_size.x)  < 0.0001: layout_size.x = 1
 			if abs(layout_size.y)  < 0.0001: layout_size.y = 1
 		elif horizontal_mode in [GGComponent.ScalingMode.ASPECT_FIT,GGComponent.ScalingMode.ASPECT_FILL]:
@@ -283,12 +283,16 @@ func _resolve_component_size( component:Node, available_size:Vector2 )->Vector2:
 		ScalingMode.SHRINK_TO_FIT:
 			if not component is GGComponent: component_size.x = component.size.x
 		ScalingMode.ASPECT_FIT:
-			var fit_x = floor( (available_size.y / component.layout_size.y) * component.layout_size.x )
-			if fit_x <= available_size.x: component_size.x = fit_x
+			if v_mode == ScalingMode.ASPECT_FILL:
+				component_size.x = floor( (available_size.y / component.layout_size.y) * component.layout_size.x )
+			else:
+				var fit_x = floor( (available_size.y / component.layout_size.y) * component.layout_size.x )
+				if fit_x <= available_size.x: component_size.x = fit_x
 		ScalingMode.ASPECT_FILL:
-			var scale_x = (available_size.x / component.layout_size.x)
-			var scale_y = (available_size.y / component.layout_size.y)
-			component_size.x = floor( max(scale_x,scale_y) * component.layout_size.x )
+			if v_mode != ScalingMode.ASPECT_FIT:
+				var scale_x = (available_size.x / component.layout_size.x)
+				var scale_y = (available_size.y / component.layout_size.y)
+				component_size.x = floor( max(scale_x,scale_y) * component.layout_size.x )
 		ScalingMode.FIXED:
 			component_size.x = component.layout_size.x
 		ScalingMode.PARAMETER:
@@ -305,12 +309,16 @@ func _resolve_component_size( component:Node, available_size:Vector2 )->Vector2:
 		ScalingMode.SHRINK_TO_FIT:
 			if not component is GGComponent: component_size.y = component.size.y
 		ScalingMode.ASPECT_FIT:
-			var fit_y = floor( (available_size.x / component.layout_size.x) * component.layout_size.y )
-			if fit_y <= available_size.y: component_size.y = fit_y
+			if h_mode == ScalingMode.ASPECT_FILL:
+				component_size.y = floor( (available_size.x / component.layout_size.x) * component.layout_size.y )
+			else:
+				var fit_y = floor( (available_size.x / component.layout_size.x) * component.layout_size.y )
+				if fit_y <= available_size.y: component_size.y = fit_y
 		ScalingMode.ASPECT_FILL:
-			var scale_x = (available_size.x / component.layout_size.x)
-			var scale_y = (available_size.y / component.layout_size.y)
-			component_size.y = floor( max(scale_x,scale_y) * component.layout_size.y )
+			if h_mode != ScalingMode.ASPECT_FIT:
+				var scale_x = (available_size.x / component.layout_size.x)
+				var scale_y = (available_size.y / component.layout_size.y)
+				component_size.y = floor( max(scale_x,scale_y) * component.layout_size.y )
 		ScalingMode.FIXED:
 			component_size.y = component.layout_size.y
 		ScalingMode.PARAMETER: